Rely on SourceSearchActionFactory
